Wire up homeserver.yaml template to recognize that it can use pgbouncer.
It will be looking for the standard POSTGRES environment variables and a USE_PGBOUNCER with a 1 for on and
a 0 for off, if it doesn't exist then it's off. Additionally, no tcp overhead because of unix sockets.

This whole experiment might be pointless, as twisted already is using a connection pool of it's own. Modify
that directly with the database(s).args.cp_[min,max] variables. Max should be 100 to simulate what we did here.