Neat tests. Remove most of them, as installing latest deps seems to add 2 minutes to setup time and only shaves 3-10 seconds off most tests.
Additionally, the new workflow is a good fit for the use case, it'll be kept for the versatility it can offer. Just
remove the 'latest-deps' options for now so it will not try and install them by default.
Remove the 'pre-run' job, as centrally locating the python and poetry version variables appears to be more trouble
than it's worth. Comment out the sytest latest-deps additions as they may be useful in the future.

Python 3.11rc2 is not ready yet. Didn't really speed up anything when it worked on the syntax, styling and lint.
Did not work on trial. Sytest and Complement didn't use Python, so there's that.

This will be
merged.