Add in support for external redis instance.
SYNAPSE_ENABLE_REDIS=1 will activate redis usage. If this isn't on, the following values will not be honored.
SYNAPSE_REDIS_HOST set to the ip of your external redis instance, defaults to 'localhost'.
SYNAPSE_REDIS_PORT set to the port number of your redis server, defaults to '6379'.
SYNAPSE_REDIS_PASSWORD set to the password of your redis server, if used. Defaults to an empty string.